added functionality to close the vault on black button in AdjustSlForm removed it from message card
