+++ features/automation/protection/common/UITypes/MultiplyVaultPillChange.ts
@@ -1,4 +1,3 @@
-
 import { OtherAction } from 'features/multiply/manage/pipes/manageMultiplyVault'
 
 export const MULTIPLY_VAULT_PILL_CHANGE_SUBJECT = 'MultiplyPillChange'
