+++ features/automation/protection/controls/AdjustSlFormControl.tsx
@@ -12,6 +12,7 @@ import { Vault } from 'blockchain/vaults'
 import { TxHelpers } from 'components/AppContext'
 import { PickCloseStateProps } from 'components/dumb/PickCloseState'
 import { SliderValuePickerProps } from 'components/dumb/SliderValuePicker'
+import { VaultViewMode } from 'components/VaultTabSwitch'
 import { CollateralPricesWithFilters } from 'features/collateralPrices/collateralPricesWithFilters'
 import { formatAmount, formatPercent } from 'helpers/formatters/format'
 import { useObservable } from 'helpers/observableHook'
@@ -31,6 +32,8 @@ import { progressStatuses } from '../common/consts/txStatues'
 import { getIsEditingProtection } from '../common/helpers'
 import { extractStopLossData, prepareTriggerData } from '../common/StopLossTriggerDataExtractor'
 import { ADD_FORM_CHANGE, AddFormChange } from '../common/UITypes/AddFormChange'
+import { MULTIPLY_VAULT_PILL_CHANGE_SUBJECT } from '../common/UITypes/MultiplyVaultPillChange'
+import { TAB_CHANGE_SUBJECT } from '../common/UITypes/TabChange'
 import { TriggersData } from '../triggers/AutomationTriggersData'
 import {
   AdjustSlFormLayout,
@@ -111,6 +114,18 @@ export function AdjustSlFormControl({
       ({ estimateGas }) => estimateGas(addAutomationBotTrigger, txData),
     )
   }, [txData])
+
+  const redirectToCloseVault = () => {
+    uiChanges.publish(TAB_CHANGE_SUBJECT, {
+      type: 'change-tab',
+      currentMode: VaultViewMode.Overview,
+    })
+
+    uiChanges.publish(MULTIPLY_VAULT_PILL_CHANGE_SUBJECT, {
+      type: 'change-multiply-pill',
+      currentStage: 'closeVault',
+    })
+  }
 
   const [gasEstimationData] = useObservable(gasEstimationData$)
 
@@ -202,56 +217,66 @@ export function AdjustSlFormControl({
   }
 
   const addTriggerConfig: RetryableLoadingButtonProps = {
-    translationKey: isStopLossEnabled ? 'update-stop-loss' : 'add-stop-loss',
-    onClick: (finishLoader: (succeded: boolean) => void) => {
-      if (tx === undefined) {
-        return
-      }
-      const txSendSuccessHandler = (transactionState: TxState<AutomationBotAddTriggerData>) => {
-        transactionStateHandler(
-          (txState) => {
-            /** TODO: This is not right place for it, this should be encapsulated,
-             * probably in similar fashion as addGasEstimation$
-             */
-            const gasUsed =
-              txState.status === TxStatus.Success ? new BigNumber(txState.receipt.gasUsed) : zero
+    translationKey: isStopLossEnabled
+      ? 'update-stop-loss'
+      : slCollRatioNearLiquidationRatio(selectedSLValue, ilkData)
+      ? 'close-vault'
+      : 'add-stop-loss',
+    onClick: slCollRatioNearLiquidationRatio(selectedSLValue, ilkData)
+      ? redirectToCloseVault
+      : (finishLoader: (succeded: boolean) => void) => {
+          if (tx === undefined) {
+            return
+          }
+          const txSendSuccessHandler = (transactionState: TxState<AutomationBotAddTriggerData>) => {
+            transactionStateHandler(
+              (txState) => {
+                /** TODO: This is not right place for it, this should be encapsulated,
+                 * probably in similar fashion as addGasEstimation$
+                 */
+                const gasUsed =
+                  txState.status === TxStatus.Success
+                    ? new BigNumber(txState.receipt.gasUsed)
+                    : zero
 
-            const effectiveGasPrice =
-              txState.status ===
-              TxStatus.Success /* Is this even correct? failed tx also have cost */
-                ? new BigNumber(txState.receipt.effectiveGasPrice)
-                : zero
+                const effectiveGasPrice =
+                  txState.status ===
+                  TxStatus.Success /* Is this even correct? failed tx also have cost */
+                    ? new BigNumber(txState.receipt.effectiveGasPrice)
+                    : zero
 
-            const totalCost =
-              !gasUsed.eq(0) && !effectiveGasPrice.eq(0)
-                ? amountFromWei(gasUsed.multipliedBy(effectiveGasPrice)).multipliedBy(tokenPrice)
-                : zero
+                const totalCost =
+                  !gasUsed.eq(0) && !effectiveGasPrice.eq(0)
+                    ? amountFromWei(gasUsed.multipliedBy(effectiveGasPrice)).multipliedBy(
+                        tokenPrice,
+                      )
+                    : zero
 
-            uiChanges.publish(ADD_FORM_CHANGE, {
-              type: 'tx-details',
-              txDetails: {
-                txHash: (txState as any).txHash,
-                txStatus: txState.status,
-                txError: txState.status === TxStatus.Error ? txState.error : undefined,
-                txCost: totalCost,
+                uiChanges.publish(ADD_FORM_CHANGE, {
+                  type: 'tx-details',
+                  txDetails: {
+                    txHash: (txState as any).txHash,
+                    txStatus: txState.status,
+                    txError: txState.status === TxStatus.Error ? txState.error : undefined,
+                    txCost: totalCost,
+                  },
+                })
               },
-            })
-          },
-          transactionState,
-          finishLoader,
-          waitForTx,
-        )
-      }
+              transactionState,
+              finishLoader,
+              waitForTx,
+            )
+          }
 
-      const sendTxErrorHandler = () => {
-        finishLoader(false)
-      }
+          const sendTxErrorHandler = () => {
+            finishLoader(false)
+          }
 
-      // TODO circular dependency waitForTx <-> txSendSuccessHandler
-      const waitForTx = tx
-        .sendWithGasEstimation(addAutomationBotTrigger, txData)
-        .subscribe(txSendSuccessHandler, sendTxErrorHandler)
-    },
+          // TODO circular dependency waitForTx <-> txSendSuccessHandler
+          const waitForTx = tx
+            .sendWithGasEstimation(addAutomationBotTrigger, txData)
+            .subscribe(txSendSuccessHandler, sendTxErrorHandler)
+        },
     isStopLossEnabled,
     isLoading: false,
     isRetry: false,
@@ -259,8 +284,7 @@ export function AdjustSlFormControl({
     disabled:
       !isOwner ||
       (!isEditing && uiState?.txDetails?.txStatus !== TxStatus.Success) ||
-      (!isEditing && !uiState?.txDetails) ||
-      slCollRatioNearLiquidationRatio(selectedSLValue, ilkData),
+      (!isEditing && !uiState?.txDetails),
   }
 
   const dynamicStopLossPrice = vault.liquidationPrice
